+++ Assets/Scripts/Enemies/Shooter/ShooterGoomba.cs
@@ -71,7 +71,7 @@ public class ShooterGoomba : ShooterCreature
         }
     }
 
-    public override void OnHitboxEvent(Hitbox hitbox, int damage)
+    public override void OnHitboxEvent(Hitbox hitbox, float damage)
     {
         base.OnHitboxEvent(hitbox, damage);
