+++ Assets/Scripts/IPhaseManager.cs
@@ -5,8 +5,8 @@ public interface IPhaseManager
     void SetDreamPhase(Nightmatrix nightmatrix);
     void SetNightmarePhase(Nightmatrix nightmatrix);
 
-    void TakeDamage(int damage);
-    int GetHealth();
-    void SetHealth(int value);
+    void TakeDamage(float damage);
+    float GetHealth();
+    void SetHealth(float value);
     void Die();
 }
