+++ Assets/Scripts/Utilities/Hitbox.cs
@@ -12,6 +12,6 @@ public class Hitbox : MonoBehaviour
     public enum Type { Hammer, Simple }
 
     public ID id;
-    public int damage = 1;
+    public float damage = 1;
     [HideInInspector] public Vector2 direction;
 }
